Move ui test utilities to its own module
