+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/AcquireTokenNetworkAbstractTest.java
@@ -2,8 +2,8 @@ package com.microsoft.identity.client.msal.automationapp;
 
 import androidx.test.core.app.ApplicationProvider;
 
-import com.microsoft.identity.client.msal.automationapp.app.IApp;
-import com.microsoft.identity.client.msal.automationapp.browser.BrowserChrome;
+import com.microsoft.identity.client.ui.automation.app.IApp;
+import com.microsoft.identity.client.ui.automation.browser.BrowserChrome;
 import com.microsoft.identity.internal.testutils.labutils.LabUserHelper;
 import com.microsoft.identity.internal.testutils.labutils.LabUserQuery;
