--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/UiResponse.java
@@ -1,6 +1,0 @@
-package com.microsoft.identity.client.msal.automationapp;
-
-public enum UiResponse {
-    ACCEPT,
-    DECLINE
-}
