--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/app/IApp.java
@@ -1,14 +1,0 @@
-package com.microsoft.identity.client.msal.automationapp.app;
-
-public interface IApp {
-
-    void install();
-
-    void launch();
-
-    void clear();
-
-    void uninstall();
-
-    void handleFirstRun();
-}
