--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/browser/IBrowser.java
@@ -1,7 +1,0 @@
-package com.microsoft.identity.client.msal.automationapp.browser;
-
-import com.microsoft.identity.client.msal.automationapp.app.IApp;
-
-public interface IBrowser extends IApp {
-
-}
