+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/TestCase833514.java
@@ -1,16 +1,16 @@
 package com.microsoft.identity.client.msal.automationapp.testpass.broker;
 
 import com.microsoft.identity.client.MultipleAccountPublicClientApplication;
-import com.microsoft.identity.client.Prompt;
 import com.microsoft.identity.client.PublicClientApplication;
 import com.microsoft.identity.client.SingleAccountPublicClientApplication;
 import com.microsoft.identity.client.exception.MsalException;
 import com.microsoft.identity.client.msal.automationapp.AcquireTokenNetworkAbstractTest;
 import com.microsoft.identity.client.msal.automationapp.R;
-import com.microsoft.identity.client.msal.automationapp.broker.BrokerAuthenticator;
-import com.microsoft.identity.client.msal.automationapp.broker.ITestBroker;
-import com.microsoft.identity.client.msal.automationapp.web.MicrosoftPromptHandler;
-import com.microsoft.identity.client.msal.automationapp.web.PromptHandlerParameters;
+import com.microsoft.identity.client.ui.automation.PromptParameter;
+import com.microsoft.identity.client.ui.automation.broker.BrokerAuthenticator;
+import com.microsoft.identity.client.ui.automation.broker.ITestBroker;
+import com.microsoft.identity.client.ui.automation.web.MicrosoftPromptHandler;
+import com.microsoft.identity.client.ui.automation.web.PromptHandlerParameters;
 import com.microsoft.identity.internal.testutils.labutils.LabConfig;
 import com.microsoft.identity.internal.testutils.labutils.LabConstants;
 import com.microsoft.identity.internal.testutils.labutils.LabUserHelper;
@@ -67,7 +67,7 @@ public class TestCase833514 extends AcquireTokenNetworkAbstractTest {
                 .sessionExpected(false)
                 .consentPageExpected(false)
                 .broker(mBroker)
-                .prompt(Prompt.SELECT_ACCOUNT)
+                .prompt(PromptParameter.SELECT_ACCOUNT)
                 .expectingNonZeroAccountsInBroker(false)
                 .build();
