+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/local/BrokerLessMsalTest.java
@@ -1,7 +1,7 @@
 package com.microsoft.identity.client.msal.automationapp.testpass.local;
 
 import com.microsoft.identity.client.msal.automationapp.AcquireTokenNetworkAbstractTest;
-import com.microsoft.identity.client.msal.automationapp.broker.ITestBroker;
+import com.microsoft.identity.client.ui.automation.broker.ITestBroker;
 
 public abstract class BrokerLessMsalTest extends AcquireTokenNetworkAbstractTest {
