+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/local/TestCase99267.java
@@ -3,11 +3,12 @@ package com.microsoft.identity.client.msal.automationapp.testpass.local;
 import com.microsoft.identity.client.AcquireTokenParameters;
 import com.microsoft.identity.client.Prompt;
 import com.microsoft.identity.client.msal.automationapp.R;
-import com.microsoft.identity.client.msal.automationapp.UiResponse;
 import com.microsoft.identity.client.msal.automationapp.interaction.InteractiveRequest;
 import com.microsoft.identity.client.msal.automationapp.interaction.OnInteractionRequired;
-import com.microsoft.identity.client.msal.automationapp.web.MicrosoftPromptHandler;
-import com.microsoft.identity.client.msal.automationapp.web.PromptHandlerParameters;
+import com.microsoft.identity.client.ui.automation.PromptParameter;
+import com.microsoft.identity.client.ui.automation.UiResponse;
+import com.microsoft.identity.client.ui.automation.web.MicrosoftPromptHandler;
+import com.microsoft.identity.client.ui.automation.web.PromptHandlerParameters;
 import com.microsoft.identity.internal.testutils.labutils.LabConfig;
 import com.microsoft.identity.internal.testutils.labutils.LabConstants;
 import com.microsoft.identity.internal.testutils.labutils.LabUserQuery;
@@ -42,7 +43,7 @@ public class TestCase99267 extends BrokerLessMsalTest {
                         final String password = LabConfig.getCurrentLabConfig().getLabUserPassword();
 
                         final PromptHandlerParameters promptHandlerParameters = PromptHandlerParameters.builder()
-                                .prompt(Prompt.SELECT_ACCOUNT)
+                                .prompt(PromptParameter.SELECT_ACCOUNT)
                                 .loginHintProvided(true)
                                 .sessionExpected(false)
                                 .consentPageExpected(true)
