--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/web/ILoginComponentHandler.java
@@ -1,22 +1,0 @@
-package com.microsoft.identity.client.msal.automationapp.web;
-
-public interface ILoginComponentHandler {
-
-    void handleEmailField(final String username);
-
-    void handlePasswordField(final String password);
-
-    void handleBackButton();
-
-    void handleNextButton();
-
-    void handleAccountPicker(final String username);
-
-    void confirmConsentPageReceived();
-
-    void acceptConsent();
-
-    void declineConsent();
-
-    void handleSpeedBump();
-}
