--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/web/IPromptHandler.java
@@ -1,6 +1,0 @@
-package com.microsoft.identity.client.msal.automationapp.web;
-
-public interface IPromptHandler {
-
-    void handlePrompt(String username, String password);
-}
